Add ToDoLists to database model

+++ backend/src/service/todoService.ts
@@ -1,5 +1,7 @@
-import { getTodoRepository } from "../orm/datasource.js"; 
-import { Todo } from "../orm/entity/todo.js"; 
+import { Repository } from 'typeorm';
+import { getTodoRepository, getTodoListRepository } from '../orm/datasource.js';
+import { Todo } from '../orm/entity/todo.js';
+import { TodoList } from '../orm/entity/todoLists.js';
 
 export default {
     async getAllTodos(): Promise<Todo[]> {
@@ -12,9 +14,15 @@ export default {
         return await todoRepository.findOneBy({ id });
     },
 
-    async createTodo(data: Partial<Todo>): Promise<Todo> {
-        const todoRepository = await getTodoRepository();
-        const todo = todoRepository.create(data);
+    async createTodo(data: Partial<Todo>, listId: number|undefined): Promise<Todo> {
+        const todoRepository: Repository<Todo> = await getTodoRepository();
+        const listRepository: Repository<TodoList> = await getTodoListRepository();
+        const todo: Todo = todoRepository.create(data);
+        if (listId) {
+            const list: TodoList = await listRepository.findOneBy({ id: listId });
+            todo.todoList = list;
+            list.todos.push(todo);
+        }
         return await todoRepository.save(todo);
     },
 
@@ -23,16 +31,23 @@ export default {
         const todo = await todoRepository.findOneBy({ id });
 
         if (!todo) {
-            return null; 
+            return null;
         }
 
-        Object.assign(todo, data); 
+        Object.assign(todo, data);
         return await todoRepository.save(todo);
     },
 
     async deleteTodo(id: number): Promise<boolean> {
         const todoRepository = await getTodoRepository();
+        const todo: Todo = await todoRepository.findOneBy({ id: id });
+        if (!todo) {
+            return false;
+        }
+        if (todo.todoList) {
+            todo.todoList.todos = todo.todoList.todos.filter(t => t.id !== id);
+        }
         const result = await todoRepository.delete(id);
-        return result.affected !== 0; 
-    }
+        return result.affected === 1;
+    },
 };
